Solved merge conflict and added some missing things

+++ server/communication.c
@@ -35,7 +35,7 @@ int replyToClient_error(unsigned int pid, int error_status) {
     return 0;
 }
 
-int replyToClient_success(unsigned int pid, size_t n_reserved_seats, unsigned int reserved_seats[]) {
+int replyToClient_success(unsigned int pid, int n_reserved_seats, unsigned int reserved_seats[]) {
     if(n_reserved_seats <= 0) {
         return -1;
     }
